pyproject.toml
[tool.poetry]
name = "arkdata"
version = "2.0.0"
description = "A program for fetching Arknights game data"
license = "AGPL-3.0-or-later"
authors = ["astral4 <88992929+astral4@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/astral4/arkdata"
repository = "https://github.com/astral4/arkdata"

[tool.poetry.dependencies]
python = "^3.12"
UnityPy = "^1.10.6"
Pillow = "^10.1.0"
pycryptodome = "^3.19.0"
pymongo = "^4.5.0"
fsb5 = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
